+++ src/components/layout/content/user/RequestSelectTop/RequestSelectTop.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const RequestSelectTop = () => {
+  return <div></div>;
+};
+
+export default RequestSelectTop;
